chore: adjust ai config
